RunControl: Fix 'Stop'
Terminate/kill current runners, so that the GUI will reflect that quite
soon and not if all runners finished.

+++ plugins/clangstaticanalyzer/clangstaticanalyzerrunner.cpp
@@ -84,8 +84,13 @@ ClangStaticAnalyzerRunner::ClangStaticAnalyzerRunner(const QString &clangExecuta
 ClangStaticAnalyzerRunner::~ClangStaticAnalyzerRunner()
 {
     const QProcess::ProcessState processState = m_process.state();
-    if (processState == QProcess::Starting || processState == QProcess::Running)
-        m_process.kill();
+    if (processState == QProcess::Starting || processState == QProcess::Running) {
+        m_process.terminate();
+        if (!m_process.waitForFinished(500)) {
+            m_process.kill();
+            m_process.waitForFinished();
+        }
+    }
 }
 
 bool ClangStaticAnalyzerRunner::run(const QString &filePath, const QStringList &compilerOptions)
